@@ -1,6 +1,6 @@
 {
   "name": "monnify",
-  "version": "1.3.1",
+  "version": "1.3.2",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -10,45 +10,65 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/common": "~9.1.6",
-    "@angular/core": "~9.1.6",
-    "@angular/forms": "~9.1.6",
-    "@angular/platform-browser": "~9.1.6",
-    "@angular/platform-browser-dynamic": "~9.1.6",
-    "@angular/router": "~9.1.6",
-    "@capacitor/core": "2.1.2",
-    "@ionic-native/core": "^5.0.7",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^5.0.0",
-    "angular-monnify": "^1.1.0",
-    "rxjs": "~6.5.1",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/common": "~12.0.1",
+    "@angular/core": "~12.0.1",
+    "@angular/fire": "^6.1.5",
+    "@angular/forms": "~12.0.1",
+    "@angular/platform-browser": "~12.0.1",
+    "@angular/platform-browser-dynamic": "~12.0.1",
+    "@angular/router": "~12.0.1",
+    "@angular/service-worker": "~12.0.1",
+    "@capacitor/core": "^3.1.2",
+    "@ionic-native/core": "^5.36.0",
+    "@ionic-native/splash-screen": "^5.33.1",
+    "@ionic-native/status-bar": "^5.33.1",
+    "@ionic/angular": "^5.5.2",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.5",
-    "@angular/cli": "~9.1.5",
-    "@angular/compiler": "~9.1.6",
-    "@angular/compiler-cli": "~9.1.6",
-    "@angular/language-service": "~9.1.6",
-    "@capacitor/cli": "2.1.2",
-    "@ionic/angular-toolkit": "^2.1.1",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/architect": ">= 0.900 < 0.1300",
+    "@angular-devkit/build-angular": "~12.0.1",
+    "@angular-eslint/builder": "~12.0.0",
+    "@angular-eslint/eslint-plugin": "~12.0.0",
+    "@angular-eslint/eslint-plugin-template": "~12.0.0",
+    "@angular-eslint/template-parser": "~12.0.0",
+    "@angular/cli": "~12.0.1",
+    "@angular/compiler": "~12.0.1",
+    "@angular/compiler-cli": "~12.0.1",
+    "@angular/language-service": "~12.0.1",
+    "@capacitor/cli": "^3.1.2",
+    "@ionic/angular-toolkit": "^4.0.0",
+    "@types/crypto-js": "^4.0.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
+    "@types/jquery": "^3.5.6",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~5.0.0",
+    "@types/uuid": "^8.3.0",
+    "@typescript-eslint/eslint-plugin": "4.16.1",
+    "@typescript-eslint/parser": "4.16.1",
+    "eslint": "^7.6.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
+    "firebase-tools": "^8.0.0 || ^9.0.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.2.2",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "jasmine-core": ">=3.8",
+    "jasmine-spec-reporter": "~5.0.0",
+    "jsonc-parser": "^3.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.3",
+    "karma-coverage": "~2.0.3",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "open": "^7.0.3",
+    "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.2.4"
   },
   "author": {
     "email": "danstevea@gmail.com",
